Move GitHub client to GitHub package
