+++ tests/unit/client/github_client_test.py
@@ -2,7 +2,7 @@ import unittest
 
 import pytest
 
-from mcmd.client import github_client
+from mcmd.github import client
 
 
 @pytest.mark.unit
@@ -12,6 +12,6 @@ class GithubTest(unittest.TestCase):
         - Import [emx_tags-only.xlsx](https://github.com/molgenis/molgenis/files/626884/emx_tags-only.xlsx) or [emx_package-only.xlsx](https://github.com/molgenis/molgenis/files/626894/emx_package-only.xlsx) using the ```Upload``` plugin
         """
 
-        ret = github_client._parse_attachment_urls(body)
+        ret = client._parse_attachment_urls(body)
         self.assertEqual(ret, ['https://github.com/molgenis/molgenis/files/626884/emx_tags-only.xlsx',
                                'https://github.com/molgenis/molgenis/files/626894/emx_package-only.xlsx'])
